refactor: generate sitemap directly from html
