--- videos.json
@@ -1,74 +1,0 @@
-[
-  {
-    "pageUrl": "https://michaelkuell.com/",
-    "playerUrl": "https://player.vimeo.com/video/1078827186",
-    "thumbnailUrl": "https://vumbnail.com/1078827186.jpg",
-    "title": "A Ronin Story",
-    "description": "",
-    "uploadDate": "",
-    "duration": null
-  },
-  {
-    "pageUrl": "https://michaelkuell.com/",
-    "playerUrl": "https://player.vimeo.com/video/1071453736",
-    "thumbnailUrl": "https://vumbnail.com/1071453736.jpg",
-    "title": "Growing - Time Is An Asset Campaign",
-    "description": "",
-    "uploadDate": "",
-    "duration": null
-  },
-  {
-    "pageUrl": "https://michaelkuell.com/",
-    "playerUrl": "https://player.vimeo.com/video/8563489",
-    "thumbnailUrl": "https://vumbnail.com/8563489.jpg",
-    "title": "Educational Multimedia Project",
-    "description": "",
-    "uploadDate": "",
-    "duration": null
-  },
-  {
-    "pageUrl": "https://michaelkuell.com/",
-    "playerUrl": "https://player.vimeo.com/video/231089330",
-    "thumbnailUrl": "https://vumbnail.com/231089330.jpg",
-    "title": "Bear Spot Farm Project",
-    "description": "",
-    "uploadDate": "",
-    "duration": null
-  },
-  {
-    "pageUrl": "https://michaelkuell.com/",
-    "playerUrl": "https://player.vimeo.com/video/1071451245",
-    "thumbnailUrl": "https://vumbnail.com/1071451245.jpg",
-    "title": "Crafting - Time Is An Asset Campaign",
-    "description": "",
-    "uploadDate": "",
-    "duration": null
-  },
-  {
-    "pageUrl": "https://michaelkuell.com/",
-    "playerUrl": "https://player.vimeo.com/video/1029449968",
-    "thumbnailUrl": "https://vumbnail.com/1029449968.jpg",
-    "title": "Like Me! by Michael Kuell",
-    "description": "",
-    "uploadDate": "",
-    "duration": null
-  },
-  {
-    "pageUrl": "https://michaelkuell.com/",
-    "playerUrl": "https://player.vimeo.com/video/366306782",
-    "thumbnailUrl": "https://vumbnail.com/366306782.jpg",
-    "title": "UTC Annual Report - UTAS",
-    "description": "",
-    "uploadDate": "",
-    "duration": null
-  },
-  {
-    "pageUrl": "https://michaelkuell.com/",
-    "playerUrl": "https://player.vimeo.com/video/156855931",
-    "thumbnailUrl": "https://vumbnail.com/156855931.jpg",
-    "title": "Pfizer Update - Sally Sussman",
-    "description": "",
-    "uploadDate": "",
-    "duration": null
-  }
-]
